+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.6.0
 
-* Add support for fetching data from NSIDC0802 from disk.
+* Add support for fetching data from NSIDC0802 v2 from disk.
 
 ## 0.5.0